+++ library_project/src/edu/umich/eac/IntNWLabels.java
@@ -2,6 +2,6 @@ package edu.umich.eac;
 
 // XXX: import these from intnw library
 public class IntNWLabels {
-    public static final long ONDEMAND = 4;
-    public static final long BACKGROUND = 8;
+    public static final int ONDEMAND = 4;
+    public static final int BACKGROUND = 8;
 }
