+++ library_project/src/edu/umich/eac/eacJNI.java
@@ -9,8 +9,8 @@
 package edu.umich.eac;
 
 class eacJNI {
-  public final static native long JNICacheFetcher_call(long jarg1, JNICacheFetcher jarg1_, long jarg2);
-  public final static native void delete_JNICacheFetcher(long jarg1);
+  public final static native int JNICacheFetcher_call(int jarg1, JNICacheFetcher jarg1_, int jarg2);
+  public final static native void delete_JNICacheFetcher(int jarg1);
 
   static {
     System.loadLibrary("eac_native");
